Improve representation of 0-byte objects
With forwarding pointers stored in an object header, the relaxation of
the RSSA type system to allow 0-byte objects, and support for
bytesPerElement == 0 in GC_arrayAllocation, normal objects and array
object elements do not have a required minimum width.  In particular, a
"unit ref" is represented by just an object header (and no content
bytes) and a "unit array" is represented by just a counter field, a
length field, and an object header (and no content bytes).

The "size" and "size2" regression tests are known to be broken by this
commit.  In particular, the size of a useless function is reported as ~8
bytes (on a 64-bit system), because the Basis Library implementation of
MLton.size is:

    fun size x =
       let
          val refOverhead =
             Int.div (HeaderWord.wordSize + ObjptrWord.wordSize, 8)
       in
          C_Size.toInt (Primitive.MLton.size (ref x)) - refOverhead
       end