Omit build date and build node from version banner
Due to embedding the build date and build node into the version banner,
a self-compile is not deterministic, which is a problem on platforms
that value reproducible builds.  In the past, the build date and build
node may have been used to triangulate the source revision and target
platform of a build, but have not been used meaningfully in recent
memory.

Closes MLton/mlton#108.