Introduce NEEDS_SIGALTSTACK_EXEC config
Rare platforms have required readable, writable, and executable memory
for an alternate signal stack.  However, modern platforms frown on
memory that is both writable and executable as a security issue.  A new
NEEDS_SIGALTSTACK_EXEC config parameter is introduced to allow a
platform to select whether or not to make the sigaltstack memory
executable.

See:
 * http://www.mlton.org/pipermail/mlton-devel/2004-December/026512.html
 * http://www.mlton.org/pipermail/mlton-devel/2008-February/030169.html
 * http://www.mlton.org/pipermail/mlton-devel/2008-February/030175.html

Closes MLton/mlton#151.