Fix parsing of numeric labels
MLton has (erroneously) allowed any integer constant denoting a positive
value to be used as numeric label; e.g.,

    val r = {001 = "a", 0xFF = "b"}

However, the Definition states

> The class Lab is extended to include the *numeric* labels `1` `2` `3`
> ..., i.e., any numeral not starting with `0`.

No other SML compiler allows a hex integer constant and only the MLKit
allows a dec integer constant with a leading zero.

With the introduction of extended literal constants, it is more
important to disallow invalid numeric labels, else there is unnecessary
ambiguity in:

    val r = {10 = "a", 1_0 = "b", 1__0 = "c"}

This commit continues to use the INT token for numeric labels, but (like
the handling of a digit for infix declarations) checks that the
representation of the INT token does not begin with 0, is not an
extended literal, is not negative, and is decimal.

An alternate strategy would be to introduce a NUMERIC token into the
scanner (and parse the integer expressions via: int = INT | NUMERIC).