VectorExps updates: Cleanup style in backend
The full change of backend.fun with respect to master is the
following:

-               Bind {dst = (var, _), isMutable, src} =>
-                  if isMutable
-                     orelse (case #operand (varInfo var) of
-                                VarOperand.Const _ => false
-                              | _ => true)
-                     then (Vector.new1
-                           (M.Statement.move {dst = varOperand var,
-                                              src = translateOperand src}))
-                  else Vector.new0 ()
+               Bind {dst = (var, _), src, ...} =>
+                  Vector.new1
+                  (M.Statement.move {dst = varOperand var,
+                                     src = translateOperand src})

This removes a superflous (and now incorrect) optimization; on a
self-compile, the "else" branch is never taken.

When a vector expression is globalized then it is quite likely that a
corresponding Vector_length of that vector is also globalized.  (The
Vector_length should be constant-propagated away, since it is known,
but that is a separate issue.)  After implementing vectors, the
following sequence may appear as part of the "globals" sequence:

    a_16: word32 array = Array_uninit(word32) (global_520)
    x_2392: unit = Array_update(word32) (a_16, global_19, global_15)
    x_2394: unit = Array_update(word32) (a_16, global_24, global_212)
    x_2396: unit = Array_update(word32) (a_16, global_521, global_213)
    x_2398: unit = Array_update(word32) (a_16, global_522, global_214)
    x_2400: unit = Array_update(word32) (a_16, global_523, global_215)
    x_2402: unit = Array_update(word32) (a_16, global_167, global_216)
    x_2404: unit = Array_update(word32) (a_16, global_524, global_94)
    x_2406: unit = Array_update(word32) (a_16, global_525, global_78)
    x_2408: unit = Array_update(word32) (a_16, global_526, global_220)
    x_2409: word32 vector = Array_toVector(word32) (a_16)
    v_0: word32 vector = x_2409
    x_655: word64 = Vector_length(word32) (v_0)

With the "optimization", the

    x_655: word64 = Vector_length(word32) (v_0)

is dropped.  Thus, x_655 is uninitialized and subsequently, when the
code compares an index to x_655 for a bounds check, it could either
erroneously fail (if the uninitialized value were 0, for example) or
erroneously succeed (if the uninitialized value were greater than the
true length of the array).

It may be that the "optimization" is a hold over from a much ealier
version of MLton, when the pipeline looked like:
  AST -> CoreML -> XML -> SXML -> CPS -> Machine
Since that time, it has evolved to:
  AST -> CoreML -> XML -> SXML -> SSA -> SSA2 -> RSSA -> Machine

CPS was a predecessor to SSA; for this discussion, the important point
is that, like SSA, it had a collection of global declarations,
including ones like:

  x_123 : word32 = 0wxF

In the generated code, we don't want to reference an integer constant
indirectly through a variable; we want it to be an immediate operand.
Since CPS was converted directly to Machine, this propagation happened
as part of the conversion.  So, the "else" branch drops such
declarations under the assumption that the propagation will happen.
In this (simple) world, that is a safe optimization.  But, it is not
necessarily an important one --- the globals are initialized once at
the beginning of the program; there would be very little overhead in
actually writing that integer constant into a global variable, even if
that global variable was never read again.

As the compiler evolved, we added the RSSA (representation SSA)
intermediate language.  In this intermediate language, we are able to
explicitly perform propagation of constants and actually drop the
corresponding declaration.  The "else" branch still exists in the RSSA
to Machine translation, but is never taken.

It seems that RSSA (and the RSSA to Machine translation) is not
performing the propagation of something like:

  x_20: Word64 = OW64 (Cast (a_0, Objptr (opt_4)), ~16): Word64

which is the explicit reading of the length field from the vector in
variable a_0.  It should be able to perform this propagation, but the
program should also compile without the propagation.  (It is also a
little more subtle; one cannot simply propagate everything like this,
because expressions like the above are also what is used to read
elements from a (mutable) array, and it wouldn't be safe to move the
reading of an array around, since the array might be modified.)

With vector expressions (and consequently vector length calculations)
lifted to globals, it is simply not true that every "simple" global is
fully propagated through the program and we shouldn't ever take the
"else" branch.  And, given the optimization of the RSSA intermediate
language, we've already explicitly propagated the ones that we can, so
we shouldn't ever need to perform the optimization at the RSSA to
Machine conversion