VectorExps updates: Use (nullary) Vector_vector instead of Array_array0Const
Replace the use of Array_array0Const primitive in the Basis Library
implementation with a (nullary) Vector_vector primitive.

The original motivation for the Array_array0Const primitive was to
share the heap space required for zero-length vectors among all
vectors (of a given type).  It was claimed that this optimization is
important, e.g., in a self-compile, where vectors are used for lots of
syntax tree elements and many of those vectors are empty.  See:
  http://www.mlton.org/pipermail/mlton-devel/2002-February/021523.html

Curiously, the full effect of this optimization has been missing for
quite some time (perhaps since the port of ConstantPropagation to the
SSA IL).  While ConstantPropagation has "globalized" the nullary
application of the "Array_array0Const" primitive, it also
simultaneously transformed it to an application of the
"Array_uninit" (previously, the "Array_array") primitive to the zero
constant.  The hash-consing of globals, meant to create exactly one
global for each distinct constant, treats "Array_uninit" primitives as
unequal (appropriately, since "Array_uninit" allocates an array with
identity (though the identity may be supressed by a subsequent
"Array_toVector")), hence each distinct "Array_array0Const" primitive
in the program remained as distinct globals.  The limited amount of
inlining prior to ConstantPropagation meant that there were typically
fewer than a dozen "copies" of the same empty vector in a program for
a given type.

As a "functional" primitive, a nullary "Vector_vector" is globalized
by ClosureConvert, but is further recognized by ConstantPropagation
and hash-consed into a unique instance for each type.