Add support for vector patterns
Much of the front-end handles vector patterns as it does list
patterns (e.g., parsing, type-checking).  However, vector patterns are
carried through to match compilation (whereas list patterns are
expanded to :: and nil patterns during defunctorize before match
compilation).  Vector patterns are handled well by the match compiler,
taking cues from both match compilation of ConApp and Tuple patterns:
a "column" of vector patterns is partitioned according to length; a
"case" on the test expression's vector length dispatches to the
correct group and a sequence of vector subscripts extracts all of the
components for subsequent pattern matching.  Non-exhaustive and
redundant pattern warnings are naturally supported.

The one design choice is how to represent a non-exhaustive pattern
match.  For example, with the input

    datatype t = A | B | C | D of t

    fun f (t: t vector) =
       case t of
          #[A, A] => B
        | #[B, B] => C
        | #[C, C] => A
        | #[D _, D _] => D (D A)

the warning is:

    Warning: z.sml 4.4.
      Case is not exhaustive.
        missing pattern: (#[] | #[_] | #[_, _, _, ...]
                          | #[A, B | C | D _]
                          | #[C, A | B | D _]
                          | #[D _, A | B | C]
                          | #[B, A | C | D _])
        in: case t of #[A, A] => B | #[B, B]   ...  _, D _] => (D (D A))

All lengths of vectors not covered by the match are included in the
non-exhaustive warning; the "#[_, _, _, ...]" is meant to indicate an
vector of length greater than or equal to three.  Technically, this is
not a pattern that the user could write (though, neither were
"or"-patterns for quite some time).  It might be interesting to
support such vector patterns (and possibly similar list patterns) in
the future.