Fix Ssa.PrePasses.breakCriticalEdges for codeMotion
To accomodate code motion, a block with multiple successors should
have a unique predecessor with a Goto transfer, which can accomodate
both new code and an addition of arguments.  This was mostly enforced,
but a Case transfer with one successor (e.g., for a datatype with one
variant) could violate.