Limit Machine IR Globals to vars used in non-main functions
Previously, every variable bound in the main function of the RSSA
IR (used to allocate globals) is allocated as a Machine IR global.
However, many variables bound in the main function are simply required
as components to compute the "real" global that is used by the rest of
the program.  For example, consider a list literal that is lifted to a
global; in the SSA and RSSA IRs, each tail is also bound as a global,
but only the final cons is referenced elsewhere in the program.

A simple walk over the non-main functions of the RSSA IR program
identifies those variables used outside of the main function and only
variables that are defined in the main function and used outside of
the main function are allocated as globals; other variables can be
allocated as Stack or Register as appropriate.

On a self-compile, globals change from:

    PRIVATE CPointer globalCPointer [3];
    PRIVATE Int8 globalInt8 [0];
    PRIVATE Int16 globalInt16 [0];
    PRIVATE Int32 globalInt32 [0];
    PRIVATE Int64 globalInt64 [0];
    PRIVATE Objptr globalObjptr [9157];
    PRIVATE Real32 globalReal32 [7];
    PRIVATE Real64 globalReal64 [9];
    PRIVATE Word8 globalWord8 [0];
    PRIVATE Word16 globalWord16 [377];
    PRIVATE Word32 globalWord32 [524];
    PRIVATE Word64 globalWord64 [625];
    PRIVATE Pointer globalObjptrNonRoot [1];

to:

    PRIVATE CPointer globalCPointer [3];
    PRIVATE Int8 globalInt8 [0];
    PRIVATE Int16 globalInt16 [0];
    PRIVATE Int32 globalInt32 [0];
    PRIVATE Int64 globalInt64 [0];
    PRIVATE Objptr globalObjptr [6180];
    PRIVATE Real32 globalReal32 [7];
    PRIVATE Real64 globalReal64 [9];
    PRIVATE Word8 globalWord8 [0];
    PRIVATE Word16 globalWord16 [80];
    PRIVATE Word32 globalWord32 [4];
    PRIVATE Word64 globalWord64 [73];
    PRIVATE Pointer globalObjptrNonRoot [1];

This saves some code size and runtime space (since these globals are
statically allocated).  Furthermore, reducing the number of global
objptrs saves some GC work, since each GC needs to iterate through all
of the global objptrs.  In general, heap space won't be saved,
since (as described above) most variables in the main function that
are not global are used to construct objects that are global objptrs,
so the objects allocated in the main function are typically retained
by their (global) containing object.