Improve CoreML to XML translation of flat record pattern matching
The translation from CoreML to XML of 'case' expressions has a special
case for record/tuple patterns of all variables, which simply
generates all of the selects as a prefix to the rhs of the match,
rather than incurring the overhead of full match compilation.

Extend that special case to record/tuple patterns of all variables or
wilds, which is a common idiom, arising from any "#f" field selection
expression or "val (x, _) = ..." declaration.