Update Emacs page
Update the example `compilation-error-regexp-alist` to support full
region error and warning messages.  Nicely, this regular expression
works with both error messages of the form:

    Error: z.sml 1.5.

and of the form:

    Error: z.sml 1.5-1.11.

However, one problem is that emacs interprets an end-line/end-column
as inclusive of the range to be highlighted, rather then being one
character beyond the range to be highlighted.