Refactor app_exp productions in ml.grm
The grammar productions for application expressions were:

    app_exp  : aexp app_exp1
             | longvid app_exp1

    app_exp1 :
             | app_exp

It seems that the empty production in app_exp1 causes ml-yacc to
associate an "app_expright" position with the left position of the
token following the app_exp, rather than with the right position of
the last token of the app_exp.  Thus, the following program:

    val i =
       case 0 of
          0 => true
        | 1 => ignore 1
        | _ => false

yields the error message:

    Error: z.sml 4.12-5.5.
      Rule results disagree.
        result:   [unit]
        previous: [bool]
        in: 0 => true | 1 => (ignore 1) | _ => false

Note that the end of the region is 5.5, corresponding to the "|" in
"| _ => false".

The grammar is refactored to:

    app_exp : aexp
            | aexp app_exp
            | longvid
            | longvid app_exp

which yields the error message:

    Error: z.sml 4.12-4.19.
      Rule results disagree.
        result:   [unit]
        previous: [bool]
        in: 0 => true | 1 => (ignore 1) | _ => false