Treat `=` as a special identifier
Check for definition and/or specification of `=` identifier along with
other special identifier checks and eliminate the `allowRebindEquals`
expert MLB annotation.

In addition to being somewhat more uniform, this fixes two minor bugs
in MLton that erroneously accepted:

    fun op= z = z

and

    signature S = sig val = : unit -> unit end

In the first case, although `val op= = fn x => x` and
`val rec op= = fn x => x` (and, generally, any use of `=` in patterns)
were correctly reported as errors, the function name in a fun
declaration was only checked against the special identifiers;
therefore, an `=` identifier was allowed to be bound.

In the second case, names in val descriptions were only checked
against the special identifiers; therefore an `=` identifier was
allowed to be specified.  Note that the Definition states "The
identifier = may not be re-bound".  Arguably, the Definition allows
`=` to be specified in a signature.  However, the use of a signature
that specifies `=` in a functor argument would constitute a binding of
`=` (although, it would not be possible to apply the functor, as it
would not be possible to define a structure that binds `=` so as to
match the signature).  Hence, we choose to treat `=` as a special
identifier and exclude it from specifications and definitions.