Treat "in: ..." context in error messages more uniformly
Try to avoid having the direct call to `Control.error` construct an
extra message of the form `seq [str "in: ", lay ()]`; rather, require
the caller to supply an approriate "context" message thunk.  One
advantage is that a caller can simply provide `fn () => empty` when
wishing to suppress the extra message.