Update signature matching error messages
This begins updating some error messages related to the module
system.  Consider the following program:

    signature SIG =
       sig
          val f: int -> string
          exception E of bool
       end
    structure Str : SIG =
       struct
          val f = Bool.toString
          val E = Fail
       end

Previously, the following error messages were generated:

    Error: z.sml 6.17-6.19.
      Variable type in structure disagrees with signature.
        variable: E
        structure: [string] -> _
        signature: [bool] -> _
    Error: z.sml 6.17-6.19.
      Variable in structure but exception in signature: E.
    Error: z.sml 6.17-6.19.
      Variable type in structure disagrees with signature.
        variable: f
        structure: [bool] -> _
        signature: [int] -> _

Now, the following error messages are generated:

    Error: z.sml 6.17-6.19.
      Value identifier in structure disagrees with signature: E.
        structure: <variable>
        structure: [string] -> _
        defn at:   z.sml 9.11-9.11
        signature: <exception>
        signature: [bool] -> _
        spec at:   z.sml 4.17-4.17
    Error: z.sml 6.17-6.19.
      Variable in structure disagrees with signature: f.
        structure: [bool] -> _
        defn at:   z.sml 8.11-8.11
        signature: [int] -> _
        spec at:   z.sml 3.11-3.11

Note that in both the previous and the current error messages, the
region associated with `Error:` is the `SIG` in
`structure Str : SIG = `.  While this is the only reasonable location
to associate with the structure/signature disagreement, in order to
fix the error message, at least one of the structure or signature must
be visited.  The new "defn at: ..." and "spec at: ..." lines link to
the definition and specification points for the identifier.  With
suitable editor support (e.g., emacs compilation-error-regexp-alist),
it is easy to visit one of the points to resolve the error.