Choose better candidates with Interface.flexibleTycons
The `Interface.flexibleTycons` function chooses candidate paths
to flexible type constructors in an interface to be realized
during a signature match.  Due to constraints, the same flexible
type constructor can be identified by multiple paths.

Consider the following program.

    structure S :
       sig
          structure T :
             sig
                datatype t = T
             end
          type u = T.t
       end =
    struct
       structure T =
          struct
             datatype t = T
          end
       type u = int
    end

Previously, the following error messages were generated:

    Error: z.sml 2.4-8.6.
      Constructor in structure disagrees with signature: T.T.
        structure: [T.t]
        defn at:   z.sml 12.23-12.23
        signature: [u]
        spec at:   z.sml 5.26-5.26
    Error: z.sml 2.4-8.6.
      Constructor in structure disagrees with signature: T.T.
        structure: [T.t]
        defn at:   z.sml 12.23-12.23
        signature: [u]
        spec at:   z.sml 5.26-5.26

This is due to choosing `u` as the candidate path to be realized.
When `u` is realized by `int`, it implies in the signature that
`T.T` is (incorrectly) a constructor of the `int` type, but in
the structure `T.T` is (correctly) a constructor of the `T.t`
type.

The choice of `u` as the candidate path to be realized for the
flexible type constructor was due to `u` being the shortest path
in the signature that identifies the flexible type constructor.

Now, the following error message is generated:

    Error: z.sml 2.4-8.6.
      Type in structure disagrees with signature: u.
        structure: [u]
        defn at:   z.sml 14.9-14.9
        signature: [T.t]
        spec at:   z.sml 7.12-7.12

This is due to changing the heuristic for choosing the candidate
path to prefer a path to a `datatype` specification for a
flexible type constructor that is a `datatype`.

Unfortunately, it isn't currently possible to distinguish a
replicating `datatype` specification from a non-replicating
`datatype` specification.  Hence, with the program:

    structure S :
       sig
          structure T :
             sig
                datatype t = T
             end
          datatype u = datatype T.t
       end =
    struct
       structure T =
          struct
             datatype t = T
          end
       type u = int
    end

the path `u` is chosen as the candidate and the following error
messages are generated:

    Error: z.sml 2.4-8.6.
      Constructor in structure disagrees with signature: T.T.
        structure: [T.t]
        defn at:   z.sml 12.23-12.23
        signature: [u]
        spec at:   z.sml 5.26-5.26
    Error: z.sml 2.4-8.6.
      Constructor in structure disagrees with signature: T.T.
        structure: [T.t]
        defn at:   z.sml 12.23-12.23
        signature: [u]
        spec at:   z.sml 5.26-5.26
    Error: z.sml 2.4-8.6.
      Type u is a datatype in signature but not in structure.
    Error: z.sml 2.4-8.6.
      Constructor in signature but not in structure: T.
        spec at:   z.sml 5.26-5.26