Rename vars and tyvars in ElaborateEnv.transparentCut.map
Rename the `sig*` variables and `'sig*` type variables to `ifc*` and
`'ifc*`, respectively.  This function aligns a structure with an
interface, triggering an action on each entity in the structure that
matches a correspondingly named entity in the interface and an action
on each entity in the interface without a correspondingly named entity
in the structure.

The motivation for the renaming is that the "interface" to which this
function is applied will shortly be extended from a single realized
signature interface to the pair of an unrealized signature interface
and a realized signature interface.