Perform transparentCut with unrealized and realized signatures
When performing a signature match (via `ElaborateEnv.transparentCut`),
propagate the unrealized signature, the paths to realized flexible
type constructors, and the realized signature.  The motivation is to
yield better diagnostics when matching a structure's types with the
signature's types, where it will help to know if the type was flexible
in the signature.