Update diagnostics for (recursive) function declarations
The primary motivation is to minimize the use of the (confusing)
"recursive use of function disagrees with its type" error message.  In
particular, it can be frustrating to add type information to a
function declaration in the hopes of identifying the problematic
recursive use, only to have the type information seemingly ignored.

For example, consider the following program:

    fun f (): int = 13
      | f (): bool = true

    fun g (x: int) = g true

    fun h x = (x + 1; h true)

    fun i x = j true
    and j y = y + 1

Previously, MLton generated the following error messages:

    Error: z.sml 2.18-2.21.
      Function clause with expression of different type.
        expression: [bool]
        previous:   [int]
        in: fun f (): int = 13 | f (): bool = true
    Error: z.sml 4.5-4.5.
      Recursive use of function disagrees with its type.
        function type: [int] -> _
        recursive use: [bool] -> _
        in: fun g (x: int) = (g true)
    Error: z.sml 6.5-6.5.
      Recursive use of function disagrees with its type.
        function type: [int] -> _
        recursive use: [bool] -> _
        in: fun h x = (x + 1 ; h true)
    Error: z.sml 9.5-9.5.
      Recursive use of function disagrees with its type.
        function type: [int] -> _
        recursive use: [bool] -> _
        in: fun i x = (j true) and j y = (y + 1)

Now, MLton generates the following error messages:

    Error: z.sml 2.11-2.14.
      Function clause with result constraint of different type.
        constraint: [bool]
        previous:   [int]
        in: fun f (): int = 13 | f (): bool = true
    Error: z.sml 4.18-4.23.
      Function applied to incorrect argument.
        expects: [int]
        but got: [bool]
        in: g true
    Error: z.sml 6.5-6.5.
      Recursive use of function disagrees with its type: h.
        function type: [int] -> _
        recursive use: [bool] -> _
        in: fun h x = (x + 1 ; h true)
    Error: z.sml 9.5-9.5.
      Recursive use of function disagrees with its type: j.
        function type: [int] -> _
        recursive use: [bool] -> _
        in: fun i x = (j true) and j y = (y + 1)

The second function exemplifies the typical case, where the type
annotations on a function are used when entering the function into the
environment, leading to type errors at the use site.  This follows the
heuristic that explicitly written information is more likely correct
than implicitly inferred information.

The third and fourth functions clarify that only explicit information
in the function patterns/result-constraint are used when entering the
function into the environment; in the absence of explicit information,
the function is given a fresh unification variable when entered into
the environment, which is unified with the type inferred for the
function code.