Introduce AstModules.WhereEquation
The AstModules.WhereEquation.t type is a `node Wrap.t` type that
carries region for the `where` equation.  In particular, the left
position of the region is the left position of the introductory
keyword.  Consider the following program:

    signature T =
       sig
          type t
          type u
       end
    structure Z : T where type t = bool
                      and type u = int =
       struct
          datatype t = T of bool
          datatype u = U of int
       end

Previously MLton generated the following error messages:

    Error: z.sml 6.15-7.34.
      Type in structure disagrees with signature (type structure): t.
        structure: [datatype] t = ...
        defn at:   z.sml 9.16-9.16
        signature: [type] t = bool
        spec at:   z.sml 3.12-3.12
        spec at:   z.sml 6.28-6.35
    Error: z.sml 6.15-7.34.
      Type in structure disagrees with signature (type structure): u.
        structure: [datatype] u = ...
        defn at:   z.sml 10.16-10.16
        signature: [type] u = int
        spec at:   z.sml 4.12-4.12
        spec at:   z.sml 7.28-7.34

Note that the second `spec at:` diagnostics have regions that
correspond to "t = bool" and "u = int"; this is in contrast to similar
diagnostics for `sharing` equations, where the regions include the
"sharing type" keywords.

Now, MLton generates the following error messages:

    Error: z.sml 6.15-7.34.
      Type in structure disagrees with signature (type structure): t.
        structure: [datatype] t = ...
        defn at:   z.sml 9.16-9.16
        signature: [type] t = bool
        spec at:   z.sml 3.12-3.12
        spec at:   z.sml 6.17-6.35
    Error: z.sml 6.15-7.34.
      Type in structure disagrees with signature (type structure): u.
        structure: [datatype] u = ...
        defn at:   z.sml 10.16-10.16
        signature: [type] u = int
        spec at:   z.sml 4.12-4.12
        spec at:   z.sml 7.19-7.34

Note that the second `spec at:` diagnostics have regions that
correspond to "where type t = bool" and "and type u = int".

One minor changes is that AstModules.Sigexp.layout will display all
where equations with `where type`, rather than `where type` for the
first equation and `and type` for the subsequent equations.  Since
`and type` for where equations is a known "wart" in the grammar (and
completely equivalent to a sequence of `where type` equations), this
seems to be an acceptable compromise.

The introduction of WhereEquation also opens the possibility of
supporting SML/NJ-style `where <longstrid> = <strexp>`
equations (although the grammar seems complicated and the semantics
are not clear).