Improve choice of tycon pretty names during signature match
Consider the following program:

    structure S :
    sig
       eqtype t
       val toStringT: t -> string
       eqtype u
       val toStringU: u -> string
    end =
    struct
       type t = real
       val toStringT = Real.toString
       datatype u = U of real
       fun toStringU (U r) = Real.toString r
    end

Previously, MLton generated the following error messages:

    Error: z.sml 2.1-7.3.
      Type in structure disagrees with signature (admits equality): t.
        structure: type t = [t]
        defn at:   z.sml 9.9-9.9
        signature: [eqtype] t
        spec at:   z.sml 3.11-3.11
    Error: z.sml 2.1-7.3.
      Type in structure disagrees with signature (admits equality): u.
        structure: datatype u = U of [t]
        defn at:   z.sml 11.13-11.13
        signature: [eqtype] u
        spec at:   z.sml 5.11-5.11
    Error: z.sml 2.1-7.3.
      Variable in structure disagrees with signature: toStringT.
        structure: [t] -> _
        defn at:   z.sml 10.8-10.16
        signature: [_sig.t] -> _
        spec at:   z.sml 4.8-4.16
    Error: z.sml 2.1-7.3.
      Variable in structure disagrees with signature: toStringU.
        structure: [u] -> _
        defn at:   z.sml 12.8-12.16
        signature: [_sig.u] -> _
        spec at:   z.sml 6.8-6.16

Note that the type "real" is referred to as "t", because tycon names
are chosen with the structure being constrained in scope.  The
"_sig.t" and "_sig.u" types are meant to indicate "in the signature of
the match", but "t" and "u" types are unqualified.

Now, MLton generates the following error messages:

    Error: z.sml 2.1-7.3.
      Type in structure disagrees with signature (admits equality): t.
        structure: type t = [real]
        defn at:   z.sml 9.9-9.9
        signature: [eqtype] t
        spec at:   z.sml 3.11-3.11
    Error: z.sml 2.1-7.3.
      Type in structure disagrees with signature (admits equality): u.
        structure: datatype u = U of [real]
        defn at:   z.sml 11.13-11.13
        signature: [eqtype] u
        spec at:   z.sml 5.11-5.11
    Error: z.sml 2.1-7.3.
      Variable in structure disagrees with signature: toStringT.
        structure: [real] -> _
        defn at:   z.sml 10.8-10.16
        signature: [_sig.t] -> _
        spec at:   z.sml 4.8-4.16
    Error: z.sml 2.1-7.3.
      Variable in structure disagrees with signature: toStringU.
        structure: [_str.u] -> _
        defn at:   z.sml 12.8-12.16
        signature: [_sig.u] -> _
        spec at:   z.sml 6.8-6.16

The essence of the implementation is to extend the heuristic for
choosing tycon names to include priorities in addition to length (in
terms of strids) of the name.  First, with lowest priority, the
realized signature is processed (with the "_sig" prefix).  Next, with
middle priority, the structure being matched is processed (with the
"_str" prefix).  Finally, with highest priority, the environment at
the point of the match is processed (with no prefix).  The effect is
that tycons that only exist in the realized signature (meaning that
they were dummy tycons introduced to accommodate a missing or
incompatible type from the structure) will have a "_sig" prefix,
tycons that only exist in the structure (e.g., datatypes declared in
the structure or types from sub-structures that are opaquely
constrained) will have a "_str" prefix, and tycons from the ambient
environment will have no prefix.  (And tycons that are not expressible
by names in scope will have a ".?" prefix on their original names, due
to the Tycon.resetLayoutPrefix.)