Add `{prefixUnset: bool}` arg to `ElaborateEnv.setTyconLayoutPretty`
Rather than choose whether or not to include a "?." prefix on tycons
that are not expressible by names in scope based on whether or not the
environment is "top-level", make the choice at the client requiring
tycon names to be set.  In particular, for error messages, use
`{prefixUnset = true}`, while for "-show-basis" and "-show-def-use",
use `{perfixUnset = false}`.