In `andalso`, `orelse`, and `e: ty`, elaborate to required type
Due to the expansion of `e1 andalso e2` into
`if e1 then false else e2`, the elaborated type of an `andalso`
expression is taken to be the elaborated type of `e2`, even if `e2` is
known not to be of type bool.  Similarly, the elaborated type of
`e: ty` is the elaborated type of `e`, even if `e` is known not to be
of type `ty`.

In these cases, force the type of the elaborated expression to be
`bool` or `ty`, respectively, which can reduce spurious subsequent
errors.