Change Interface.TypeStr invariant
Previously, when creating a type structure from a scheme that was an
eta-expansion of a type constructor, the type structure was recorded
as a Tycon, rather than as a Scheme.  But, this makes it difficult to
determine if the specification was abstract (`type t`) or a
definition (`type t = int`).

Shift the eta-check to Interface.TypeStr.toTyconOpt.