Partially revert 0771a0a (delay remove of duplicate spec regions)
After commit 0771a0a, a self-compile was impossible due to an
explosion in memory use and time.  In particular, elaboration did not
proceed past the `sharing Ssa.Atoms = Ssa2.Atoms` of
`ssa/ssa-to-ssa2.sig` (the first such sharing of `Atoms` in the
elaboration of `mlton.mlb`).  The pervasive use of structure sharing
in the MLton sources combined with the necessary copying of `!ss'` in
`sharingSpec :: (!ss' @ !ss)` lead to excessively large lists of spec
regions.